+++ src/cmfile.jl
@@ -12,7 +12,10 @@ component that will be used in any interpolated values (using the `CommonMark`
 
 When `Revise.jl` is active and tracking the package which contains a
 `@cm_component` definition updates to the source markdown file are tracked and
-passed to `Revise` to perform code revision.
+will cause the rendered component to be updated without the need to redefine
+the component manually. If `Revise.jl` is not active then the component
+definition's markdown AST is generated at compile-time and reused on each
+render.
 """
 macro cm_component(expr)
     name, parameters, path = if MacroTools.@capture(expr, name_(; parameters__) = path_)
@@ -23,77 +26,63 @@ macro cm_component(expr)
         error("invalid `@cm_component` synctax. Must be: name(; params...) = path")
     end
 
-    dir = dirname(String(__source__.file))
-    return quote
-        $(Base).include_dependency(joinpath($dir, $(esc(path))))
-        $(Base).include(
-            $(__module__),
-            $(CMFile)(
-                joinpath($(dir), $(esc(path))),
-                $(__module__),
-                $(QuoteNode(name)),
-                $(QuoteNode(parameters)),
-            ),
-        )
-        $(Expr(
-            :toplevel,
-            Expr(
-                :module,
-                true,
-                esc(gensym("markdown_component_watcher")),
-                Expr(
-                    :block,
-                    esc(
-                        :(
-                            __init__() = $(HypertextTemplates)._includet(
-                                $(__module__),
-                                $(CMFile)(
-                                    joinpath($(dir), $(path)),
-                                    $(__module__),
-                                    $(QuoteNode(name)),
-                                    $(QuoteNode(parameters)),
-                                ),
-                            )
-                        ),
-                    ),
-                ),
-            ),
-        ))
-    end
+    # Handle direct REPL usage where `__source__` is not defined.
+    dir = isnothing(__source__.file) ? pwd() : dirname(String(__source__.file))
+
+    path_const = gensym(Symbol(name, :_path))
+    text_const = gensym(Symbol(name, :_text))
+    mod_const = gensym(Symbol(name, :_mod))
+    gen_func_name = gensym(Symbol(name, :_gen))
+
+    parameter_names = _extract_parameter_name.(parameters)
+
+    return esc(
+        quote
+            const $(path_const) = joinpath($dir, $(path))
+            const $(text_const) = Symbol(read($path_const, String))
+            const $(mod_const) = @__MODULE__
+
+            # Changes to the markdown file linked should cause the pkgimage to
+            # be invalidated since we store in `text_const` the content of the
+            # file.
+            $(Base).include_dependency($path_const)
+
+            # This allows us to cache the markdown AST generated from the
+            # `text`. `_parse_cm_content` returns an `Expr` that contains a
+            # markdown AST which may contain references to the parameters
+            # provided to this generated function. At runtime those references
+            # will be updated with the values of the individual parameters.
+            @generated function $(gen_func_name)(
+                ::Val{text};
+                $(parameters...),
+            ) where {text}
+                return $(HypertextTemplates)._parse_cm_content(
+                    $mod_const,
+                    text,
+                    $path_const,
+                )
+            end
+
+            $(HypertextTemplates).@component function $(name)(; $(parameters...))
+                ast = if $(_is_revise_loaded)()
+                    text = Symbol(read($path_const, String))
+                    # This results in a dynamic dispatch since `text` is a runtime value.
+                    $(gen_func_name)(Val{text}(); $(parameter_names...))
+                else
+                    # Ideally this should be fully inferrable since
+                    # `text_const` is a global constant.
+                    $(gen_func_name)(Val{$text_const}(); $(parameter_names...))
+                end
+                $(HypertextTemplates).@text ast
+            end
+        end,
+    )
 end
 
-struct CMFile
-    file::String
-    mod::Module
-    name::Union{Expr,Symbol}
-    parameters::Vector
-end
+# This interface function is extended by the `HypertextTemplatesCommonMarkExt`
+# extension module. See that file for the real definition.
+_parse_cm_content(mod, text, path) = error("`@cm_component` needs `CommonMark.jl` to work.")
 
-Base.String(vaf::CMFile) = vaf.file
-Base.convert(::Type{String}, file::CMFile) = String(file)
-Base.abspath(file::CMFile) =
-    CMFile(Base.abspath(file.file), file.mod, file.name, file.parameters)
-Base.isfile(file::CMFile) = Base.isfile(file.file)
-Base.isabspath(file::CMFile) = Base.isabspath(file.file)
-Base.findfirst(str::String, file::CMFile) = Base.findfirst(str, file.file)
-Base.joinpath(str::String, file::CMFile) =
-    CMFile(Base.joinpath(str, file.file), file.mod, file.name, file.parameters)
-Base.normpath(file::CMFile) =
-    CMFile(Base.normpath(file.file), file.mod, file.name, file.parameters)
-
-function Base.include(mod::Module, file::CMFile)
-    ex = cm_file_expr(mod, file)
-    Core.eval(mod, ex)
-    return nothing
-end
-
-function cm_file_expr(mod::Module, file::CMFile)
-    _cm_file_expr(mod, file, nothing)
-end
-
-function _cm_file_expr(::Module, ::CMFile, ::Any)
-    error("`CommonMark.jl` must be imported to use the `@cm` macro.")
-end
-
-_includet(mod::Module, cmfile::CMFile) = _includet(mod, cmfile, nothing)
-_includet(mod::Module, cmfile::CMFile, ::Any) = Base.include(mod, cmfile)
+_extract_parameter_name(param::Symbol) = param
+_extract_parameter_name(param::Expr) =
+    Meta.isexpr(param, :kw, 2) ? param.args[1] : error("invalid parameter syntax: $param")