+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
 - Add documentation for `@__slot__` and `$` shorthand `@text` syntax [#37]
+- Fix broken `@cm_component` macro caused by changes to `Revise` APIs [#37]
 
 ## [v2.2.2] - 2025-01-08